* src/*/*.c: Added many casts to make code more 64bit-safe.

+++ src/smooth/ftgrays.c
@@ -1037,7 +1037,7 @@
 
     for (;;)
     {
-      int    len = limit - base;
+      int    len = (int)( limit - base );
       PCell  i, j, pivot;
